Add new greetings for the scale armor crafter
Now the guy that makes RDSM will change his intro line
depending on the state of his proposal.
I had to increase the greeting array size to fit the new ones,
and it broke MMExtension, so the change turned out to be complicated.
Also in this commit: make it more clear that RDSM is a plate armor.