Combine Splitter damage messages
Hitting with Splitter triggers both a weapon attack and a fireball,
and the original combined damage messages code didn't account for this.
Now, the combined damage and kills of both attacks are printed.
However, the halved armor message will be suppressed in most cases.
Also fixes a bug that miscounted killed monsters sometimes.