Fix: prevent worst items to be rolled on high tlvl
The function for generating a random item of specified class and tlvl
had a off-by-one error that sometimes caused the first item in the class
to be chosen, even if its probability on specified tlvl is 0.
The initial attempt to fix this had resulted in crashes, as apparently
sometimes the game calls this function with no eligible items, and the
bug prevents an out-of-bounds read by enabling the said first item.
It should be OK now, although this all smells of another bug.