Treat equipped spears as 2H weapons below Master
Wielded spears at Normal or Expert skill used to trigger an awkward
special case that forbade swapping an offhand weapon or shield with them
(even if the main hand was empty) and also forbade swapping the spear
with any shield, sword, or dagger (even a 2H sword).  This was
inconveniently restrictive, so now below-Master spears are just treated
as 2H weapons: you cannot equip them if both hands are occupied,
but otherwise any swaps are possible.