Redesign the main questline
This commit rearranges the story quests, especially the Bracada/Deyja
ones.  Firstly, the player no longer has to wait 4 months after saving
dwarves for the story to progress.  As long as all Erathia/Avlee quests
are completed on time, the party is almost immediately given a
"probatory access" to both Celeste and the Pit.  This special
circumstance allows maintaining friendly relations with both parties for
a short while, until the altar pieces quest which of course can only be
completed for one side -- and the other one will then consider you a
traitor, which will cement the path choice.  You can also reject
alliance with one of the kings immediately, which gives you a bonus
later on.  So far you cannot initiate promotion quests during the
probatory stage, but that's for next time.  Took me too long already!

There's also a number of other incidental changes, such as a hack for
extending npctext.txt, or reworking the human-elf war winner logic.
I also feel I haven't playtested it all thoroughly enough.  We'll see.