Change Blaster GM bonus
Now that Blaster skill is limited by class, making it less bland by
adding nice perks is an option.  E and M already have some (which
weren't, but are now, listed in skill description), and as of this
commit, the ultimate GM perk is the ability to instakill the target.

There are some caveats.  The chance to trigger it is just skill/200,
monsters can resist vs. Magic, and it has a (mon hp)/(mon max hp) chance
of failing.  Also a successfull eradication leaves no corpse to loot.
Overall, it's powerful, but chancy, and all three of the classes that
can learn it also have (or will have) other good ranged options.

The animation for the perk is the monster shrinking away into an SFX
(used Berserk SFX here, need to draw a proper black hole/vortex later).
I wanted the monsters to turn red and disappear, but turns out the
engine does not support that; the final version is also cool, though.

Also in this commit: fixed blaster shots not carrying item information
(but did that even break anything?), and I also added some cheaty
playtesting features (not obtainable ingame).