Make GM teacher demands more varied
Now every GM teacher has a special requirement in addition to the usual
skill and gold.  For most, it's just some "practical experience" -- that
is, a certain number of successfull uses of the skill.  Some check
various stats or achievements.  A couple teachers even give mini-quests
when you try to learn the skill for the first time.

This commit took a while, but I'm basically done with skills now!  While
on it, I also changed some awkward wording in vanilla teacher text, and
refactored some code, notably the skill hint coloring.