Update to MM7Patch v2.5
Since the latest version of the patch has included some of my bugfixes,
they now need to be removed from the mod.  OTOH, some of the fixes don't
belong in the mod, so I need to turn them off (e.g. Light Bolt and
undead).  Some of the patch's new hooks also conflict with mine, and a
feature which "ungenerated" unfound artifacts caused crashes on entering
my Harmondale map, so I had to redo it.  Also Energy damage is now 12
instead of 9 (as per the patch); this changes some pre-placed monsters.
Finally, I removed the nearly-useless Sturdy enchantment, saving myself
the effort to adapt it to the new hooks.  I also added an enum for the
new global.txt strings, fixed a very wrong Hammerhands description, and
updated the UI.txt drop-in that apparently was slightly incorrect.
Oh, and I also fixed weapon elemental damage broken by 7db4a208.