Add a trident that shoots lightning
The next artifact is the Storm Trident, which deals bonus Shock damage,
grants shock immunity and water walking, and uniquely adds Lightning
Bolt to the spellbook (for as long as it's wielded) that can be cast for
free and uses Spear skill instead of Air.  So kinda like a wand in
effect, but without charges and affected by a skill.  Thus, Knights can
have a quick spell now.  Neat!  I used like 7 hooks to make the spell
display correctly despite technically not being known; in retrospect,
I should've just added it to known spells directly.  Oh well!