Add an artifact robe
Now that robes are coded in, I can introduce a planned artifact one:
Ellinger's Robe that adds +2 to all magic schools (stacks with other
magic boosters!), Preservation (like the buff), and immunity to weakness
-- I felt the need to let the latter extend even to "unresistable"
weakness, like from lack of sleep, as this is the kind encountered 90%
of the time.  Like with ordinary robes, the graphics are not done yet.