Rehaul the Temple in a Bottle
Now instead of being an entrance to the credits dungeon, it teleports
the party to an actual temple, and exiting it will restore the previous
location in the dungeon where it was last used.  The temple can heal as
usual, but it's "dark" so it turns dead PCs into zombies.  Also, the
bottle doesn't have a reputation counter, so instead donations grant
Pain Reflection at any rep level.  The level also has four chests, and
Invisibility and Lloyd's Beacon are forbidden there to prevent abuse.
The bottle item can now be found as a random artifact; it also
still appears in Shoals.  Later I'll fix that so there can only
be one per game.  Also in this commit: I lumped all the mod-specific
savegame data in a single struct, except for Walls of Mist barrels.