Add an artifact lightsaber
No, really.  Sword of Light is a Good-aligned relic that attacks with
energy, boosts light magic, and looks like a sword handle when not
equipped.  (That said, the graphics are temporary, and I'll need to
commission proper ones later.)  Also, I made daggers have 2x to-hit
bonus, and SoL also gets this property as a special exception.