Add some lower-level magic-enchancing boni
Since the main draw of the robes and redone staves is their spellcasting
bonus, but the vanilla "of X magic" boni only show up in tlvl 5-6, we
need something less powerful that can appear on low-level items as well.
So, this commit adds seven enchantments that each add +2 to one
elemental or Ego magic school, plus +10 to a related resistance for
flavor (+5 AC for Earth).  They can appear on both staves and robes, and
more rarely on amulets and crowns, but never on rings.  The skill bonus
is, naturally, not cumulative with the stronger vanilla enchantment.

Additionally, I replaced the duplicate Clanker's Amulet in the School of
Sorcery with a robe of water magic and a staff of air magic (both are
specitems), and Faerie Ring has been downgraded to the new +2 Air bonus,
as it was too powerful for how easy it's to obtain.