Add a high-damage artifact mace
Both existing mace-class artifacts mostly boosted magic, which is nice,
but perhaps some Paladins just want to crush skulls, and that niche was
empty in vanilla.  So here's an artifact mace that's made purely to
boost damage: Clover.  It adds a chance to double the total damage of an
attack equal to luck bonus %.  Note that it's much more powerful than
ordinary double damage weapons, as it also doubles skill, spell and stat
boni, as well as any offhand weapon damage.  Also note that, as always,
multiple sources of double damage do not stack.  Nevertheless, it can be
a strong choice.  At moderate values of 100 base damage and 200 luck, it
adds 20 damage on average.  At extremes of 200 damage and 500 luck, it'd
add as much as 60 damage.  Clover also adds +50 to luck, which helps.

The image for Clover is temporary, as I want it to look like a clover.
Also in this commit: removed the last limit on total artifact count.