Add a damaging relic hammer
Another high-damage mace-class weapon, just to give the player a choice.
Flattener is a quasi-two-handed hammer, in that it can be wielded in one
hand, but this penalizes melee recovery, so 2H is preferable (a bit like
a spear).  In addition to high damage, it has a 10% chance on hit to
cast Mass Distortion with power equal to wielder's Mace skill.  It's
also spectral, and MD damage is dual-elemental as well (physical/magic).

Also here: fixed Spectral weapons vs. monster Day of Protection buff.