Add a relic that lowers spell recovery
Let's throw the casters a bone for a change.  Eloquence Talisman is a
relic amulet that lowers spellcasting recovery by 25% -- i.e. with it,
a PC can cast 4 spells instead of 3 in the same time period.  This
directly increases DPS, but also SP consumption, and the relic further
drains 1 SP per 5 minutes.  Still, sorcerers and the like are gonna find
this very useful.  The relic also increases Merchant, mostly as fluff.