Add a magic-item-enhancing belt and tweak Int/Per
First, this commit makes Intellect and Personality useful for all
classes as the former now affects wand power (it's now 5 + half of Int
rating, rounded down) and the latter, merchant prices (Per rating is
now added to reputation, but the 7% bonus to Merchant is removed).

Second, I added a new artifact, the Gadgeteer's Belt, that makes wands
and scrolls cast at GM and improves their spell power by 5 (10 for
Thieves).  It also increases the power of all non-black potions used on
the wearer by 25% (50% for Thieves).  Fractions are usually retained.
The graphics for the new artifact are temporary, will be original later.