Add gloves that sometimes create reagents
The last new artifact is Gardener's Gloves.  They boost Earth magic;
more importantly, damaging a monster while wearing these gloves causes
it to sometimes drop a reagent on death.  The chance is 15% for Rangers
and Druids and 5% for all other classes.  Monsters that already drop a
reagent (e.g. trolls) are unaffected.  Power of the reagent depends on
the monster's level, but color is random.  Also, reanimated or
eradicated monsters don't ever drop reagents anymore as of this commit.