Tweak guaranteed artifacts
Now artifacts that always appear in some chests (Clanker's and Zod) will
correctly be registered as found, so duplicates won't be generated.
It's similar to MM7Patch's optional feature, but the difference is, such
artifacts won't count towards the limit of 13 found randomly.

I also added more artifact chests: obelisk reward now contains 4 random
artifacts (and the gender items are moved to the arcomage chest for
now).  One of the artifacts will be Oghma Infinium, unless it was found
before, in which case it's replaced by another random artifact.  The
Shoals chest similarly contains the temple in a bottle, or a random
artifact if it was found before.  So, 20 artifacts can be found in game.

I also made sure that the obelisk chest doesn't disappear until you
leave the map (in case some artifacts didn't fit in on the first open),
and I doubled its non-artifact loot (money, ore, reagents).  Also, all
artifact chests won't respawn their artifacts after being looted once
(the 0x45 command added by the patch was very helpful here).  Finally,
now that there can only be one temple in a bottle, I made it disappear
when inside the temple location and reappear on leaving it.