Various temporary level changes
First, this commit adds one more black potion, that gives a bonus to
character level until next rest.  The bonus is a percentage of base
level, so it can be powerful in the late game.  This is my final effort
to redeem GM alchemy, as some people still complained about its utility.

The level bonus can actually be very useful, because of how HP and SP is
calculated in the mod, but I also allowed it to affect skill values
slightly.  While at it, I standartised how Disarm and Perception boni
work - they weren't multiplied by mastery.  Learning still isn't, but
changing it would make related NPCs OP, so I just upped its level bonus.

I also re-enabled the MM6 "of Power" enchantment, which turned out to be
non-stackable, but that's fine by me.  It appears on robes, crowns, and
amulets only, and has been nerfed to +3.  Maybe I'll tweak it later.

Finally, I changed the type of every recipe to just "Recipe".