Add resistance penetration to weapon debuffs
I've meant to do this for a while.  Now skill*2 is added to resistance
penetration when a weapon triggers a debuff, just like with a proper
spell.  Affected are vanilla skill perks, cursed weapons, instadeath
effects, and the new artifacts.  I had to hardcode the address for
mm7patch's GM axe fix function, which will likely change past v2.5.7,
but at least a wrong address will not break much in my code.