Add an alternative casting mode
As requested by Eksekk, the player can now ctrl-click on some spells in
the spellbook menu to apply them to a multiple targets at a time.  For
single-PC buffs such as Regeneration and some low-rank spell versions,
it's simply cast on every PC.  For weapon-enchanting spells, they're
cast on every equipped melee weapon that can be temp-enchanted with this
spell, or every ranged weapon if there's none.  (Note: at GM, this can
result in a permanent enchantment, but it's not the intended use.)
In both cases SP cost and recovery is multiplied by the number of
targets, and targets may be limited by SP available.  This ability may
not be used near hostile monsters.  Ctrl-click is now also used to
switch off Immolation and GM Wizard's Eye, which is still free.