Tweak Enchant Item, Fire Aura, and knives
I removed the 10% * (10 - skill) chance of failure from Enchant Item, as
it was MM6 legacy made pointless by the fact that Master rank already
requires 8 skill (7 in vanilla), which automatically gives 80% success
chance.  Instead, high Water skill decreases the lowest base price at
which an item can be enchanted -- it's now 250 + 4 * skill for weapons
and twice that for other items.  For non-weapons, skill below 7 would
make this value higher than vanilla 450, but, again, outside of higher
difficulties it's impossible.  As such, it's purely an improvement.

Related, I patched the well-known Fire Aura exploit that allowed nuking
the economy by enchanting every throwaway weapon with a 2000-gold-priced
"of Infernos" bonus.  Now GM Fire Aura applies an unique version of it,
that only multiplies the weapon cost by 2 (like both Spectral and
Vampiric), but otherwise works the same.  So it's much less outrageous.

Also, while auditing the item prices (to gauge the above EI values), I
decided to cut the price of +0 and +3 throwing knives by a third, so
that the value in items.txt is no longer the min value for shop-sold
knives, but an average one.  And while testing THAT, I noticed the
Emerald Isle shop never sells knives, so now it does occasionally.