Rebalance some spells
Immolation and Invisibility now have a duration of 5/15 min/skill on
M/GM.  For the former it's an improvement (as 1 minute was ridiculously
short and could only be explained by the lack of an off-switch which I
added), for the later it's a nerf which it definitely needs.
BTW, opening chests now breaks infisibility, which is too a nerf.

Also, instead of slightly lowered recovery, Sparks and Poison Spray at
higher ranks now have significantly increased SP cost.  Also a
much-needed nerf, as at M or GM they were stupid cheap for their power.
I left Shrapnel alone, as 30 is OK and Dark should be powerful anyway.