Redo lenient alchemy
I still want druids to have an advantage alchemy-wise, so now if a
potion is successfully brewn without the proper skill, it has 25% lower
power, and to patch an obvious loophole, catalysts mixed with it are
also considered 25% less powerful.  So, while a four-knight party can
still dabble in alchemy, a druid (or an assassin) is now desirable, as
many black potions are power-sensitive (especially "pure stat" ones).