Tweak Rangers and Monks a little
As an afterthought, I've added GM Body, which is very useful at this
level, to someone besides Clerics and Druids, namely (Light) Rangers.
Now the Self school that they do not have is Spirit, and to balance it
a bit, Monks can now learn Spirit but not Mind.  Feels more fitting to
what I imagine these classes to be, too.  I also replaced Mind Blast on
NPC monks with Flying Fist, which is much more thematic as well.