Fix Nighon volcano reagents
First, there were five rubies on the top of volcano, but these were gem
rubies, not reagents, and for some reason this have caused them not to
appear in the game.  Now these are reagents and they show properly.

Second, some reagents, including the newly fixed rubies, were placed
inside the caldera, which was incredibly inconvenient (and probably
unintended), so I moved them a bit to the sides and now they're on top.