Fix a secret chest in Tularean
For some reason it was full of fixed items, and among other things it
contained a 0 power pure black potion, which is a problem in my mod, so
now it's 30 power.  Also there was a 0/0 charge wand which was recharged
to full on open (?) -- for now I changed it to 20/27, but I'll need to
fix it later.  And if I ever get to the map max treasure level rehaul,
I'll want to make this chest out-of-depth with random items instead.