Don't smile when evt.Sub an unset or hidden QBit
A small bug that was bothering me: the subtract QBit gamescript command
was triggering a smile animation unconditionally, even if the bit was
already unset or if it had no associated journal message, thus a PC
would often smile without apparent (to the player) reason.  Now the
smile only triggers if the journal is updated, just like with adding.