Fix garbage recovery on game reload
A peculiar bug that eluded me for some time, but I got it now.
Basically, in turn-based more there's an action queue that determines
who gets to act when, and it's used to calculate the recovery values
when TB is exited.  If you save the game in TB, the TB flag is saved
too, and upon reload the game will exit TB and recalculate recovery from
the queue.  But the latter isn't saved properly!  Consequently it'll
often contain garbage (usually from the last combat before reload) and
the ensuing recovery will also be wrong.  The solution is to manually
turn the TB flag off before it can be checked, and recalculate recovery
from the saved values (which need to be multiplied by 1024/225 first).