Limit Ethric's and Reaper drain effects to combat
Since Grim Reaper is only useful in combat, a sensible player would
unequip it while outside dungeons to avoid the constant SP drain.
This commit simplifies this, limiting the SP drain to unsafe locations
(yellow/red danger gem) which has basically the same overall effect.
The same is also done to Ethric's Staff HP drain; while its affects
(notably the Meditation bonus) also apply outside of combat, it's a
necessity, as the increased drain is now prone to kill the wielder while
resting, which makes the relic a liability.  So, to let the player make
use of the increased SP pool (outside of temples), the HP drain has to
be limited.  Let's just assume that only the Dark bonus requires the HP.