Reset the more information hireling flag
The flag that remembers whether the "more information" hireling dialog
option was pressed wasn't reset between two hireling dialogues.  While
this hadn't actually caused the additional information to be displayed
right at the start, switching to it would require two mouse clicks
instead of one if the flag was left set.  Now the flag is reset on exit.