A couple tweaks to turning spells off
First, Wizard Eye and Immolation can now be turned off even with 0 SP.
This never cost any SP to begin with, so it's more of a bugfix.

Second, GM Wizard Eye can now be turned off even if the current PC is
not an Air GM, e.g. if the buff was obtained via the Gadgeteer's Belt.