Tweak temporary level bonus to Learning
Upon consideration, temporary levels no longer grant an increased bonus
to Learning to compensate for the bonus not scaling with rank.  It was
kinda ugly, prone to overflow with higher skill values (60 is still the
hard cap), and it makes sense that artifical experience is not a great
help for gaining real one (although apparently it still helps a bit).