Rehaul learning racial skills ingame
As mentioned in 63c0e71, there's only one situation wherein a racial
skill is not mandatory given at start: dwarven monks and the axe skill.
As such, this commit removes the myriad hooks and checks for all the
shops and skills and only leaves a check for dwarf monks and axe in
weapon shops.  If the situation ever changes, this can be reverted.