Fix temple in a bottle and Shoals
A rather obvious exploit I only thought of now: the player could use the
temple in a bottle in Shoals, remove wetsuits, re-equip, and return to
Shoals fully armed.  The bottle can even be found there!  Now the temple
exit checks for wetsuits, as long as it was last used in that region.