Add a spoiler describing new artifacts
Actually I wrote this almost a month ago, but I was busy with work all
July.  The new files describe how the new artifacts and relics work in
detail (in-game descriptions are too brief sometimes).  I probably won't
include them in the zip, instead there's a github link in the changelog.