Display crit chance and average DPR
Now critical hit chance is conveniently displayed in the right-click
hint for the PC's damage (melee or ranged) in the stats window.
Also I've added an estimate of average damage/round there, factoring in
damage, recovery, and crit chance.  It doesn't account for elemental
damage or really anything else, so it's not terribly precise.