Fix game crashing when a cursed monster attacks
My mod uses an internal variable to remember who blocked the hit, which
is not always the hit player if Master Shield perk triggers, and later
that player gets armor skill training (for GM teacher purposes).  But a
cursed monster missing due to its curse would not set that variable,
leading to crashes if it was zero.  Now it's explicitly set to zero
(because it was specifically a miss and not a block), and the training
code checks for that zero, avoiding the crash.