Make recovery bonuses multiplicative
During playtesting it turned out that the lowered recovery limit made
melee unreasonably strong in the late game, as with a good build it's
possible to reach any positive number, be it 30 or 20 or 10.  Now, I
could raise it back to 30, but that would make many recovery factors
meaningless once the limit is reached, so this commit instead nerfs
recovery bonuses in the late game by making them multiplicative.

How this works: all armor etc. penalties are first added to base
recovery.  Then for each formerly X recovery bonus, the sum is
multiplied by (100% - X%).  This affects early game little, but as
bonuses pile up, they start to interfere with each other, so it actually
takes some effort to reach 30, and the theoretical limit of (now) 10 is
really only obtainable by Knights who still get an enormous boost from
GM Armsmaster (each individual bonus is still additive with itself!)

While I hadn't meant to nerf ranged combat too, it's still affected for
generality, so I also improved bow recovery to 80 as a compensation.
Blasters are fine as is, throwing knives hopefully are too.  We'll see.