Tweak some GM teacher requirements
Most teachers who required enough uses of skill were too lenient.  The
minimum numbers of uses are now raised according to the playtesting
data: you have to actively use the skill for a significant portion of
the game to pass now.  On the contrary, most magic GMs were too harsh
with their 50 resistance requirement, so it's lowered to 40 now.  I did
fix the exploit that made potion immunities eligible for this, though.