Some probatory access tweaks
For some reason I've initially made it so probatory access could be
granted at any time until you install a new judge, so it wasn't really
timed as the player could just delay that quest forever.  Now it only
happens if you visit the castle with all the quests done BEFORE the
timer for judge's death runs out.  Makes more sense this way.

I've also added a messenger event to Erathia and Tularean outdoors that
asks you to visit the castle if you meet the conditions above, as it's
admittedly non-obvious and now even missable.  Since the last condition
is the trumpet quest which is completed in one of the three castles,
this should be enough in most circumstances.