Rework the debuff resistance hook
Specifically, I got rid of the hardcoded address for a specific MM7Patch
function which would have to be updated with every patch version.  Now
my code checks edx for a popped return address, which is constant.  In
theory, nothing forbids edx from being equal to it in other
circumstances, but it's extremely unlikely, so good enough for me.