Add a hint dialog for graverobbing
Robbing a sarcophagus in Barrows lowers reputation since 2.0, but the
connection turned out to be non-obvious, and playtesters have been
surprised why is everything so expensive in Stone City.  In light of
this, I've added a conversation topic to two dwarven NPCs that mentions
the reputation loss in advance (or just grumbles if you already did it).