Nerf bow recovery again
A follow-up to 434a5b4: on second thought, 80 base recovery for bows is
too much of a boost -- not because they would be too strong by
themselves, but because they overshadow knives, whose primary advantage
was their speed which suffered after that commit.  Ideally I'd want to
buff knives instead, but I need more playtesting for that, and I want to
release already.  So, as a patch, raise bow recovery to 90.  This does
make them slower than in vanilla for well-developed late-game parties,
but they have blasters anyway!  Knives are now 50% faster, which is OK.