Let Learning affect quest XP
While the skill's description claims it increases all learned
experience, it actually only works on XP from killed enemies, while the
other major source of XP -- quests and other scripted events -- is
exempt.  This is arguably a bug, and it makes Learning less useful than
it appears, especially early on.  So, let's try amending this omission!