Rehaul Cure Disease
In recognition of the fact that Body needs a strong single-target heal
now that tanks are a thing, and also that a Master-level spell that
copies an orange potion is lame, Cure Disease now also heals HP
(12/skill).  It's also renamed to Elixir of Life in order to make sense
(the bottle icon helps), and cures zombie and drunk because why not.