Buff Hammerhands
It has come to my attention that Hammerhands is borderline useless:
obviously, it hardly matters without a monk in the party, but even monks
deal so much damage by the midgame that it's a drop in the bucket.  As
such, the buff now grants double the skill effect to unarmed attacks,
and half the skill effect to all other melee attacks.  Hopefully this
will make it at least somewhat justificable.