Rework Spear and Sword GM perks again
Now it's something similar to Dodging, only if melee.  Each hand with an
applicable weapon has a skill % chance to nullify a melee attack against
the PC.  In the endgame, when AC is high and monsters are high-level,
the direct chance to avoid damage is more valuable than a modest AC
bonus.  I'd say it's 40% or so more effective than the v3.0 improved
skill perk, and greatly more effective than vanilla.  It will also make
improving weapon skills more enticing, as it scales better than AC.