Make betrayed royalty pay you money
It never really made sense, from storytelling perspective, how if you
presented fake spy or plans during the human/elf war quests, the king or
queen had the exact same reaction as when you completed the quest
properly, except for paying you.  Sure, you'd be paid in the other
castle, and double money for less effort is bad, but that's balance, not
story!  So now the betrayed party pays you the same 5000 gold, and you
get 2000 more from the other one.  That's still 40% more gold, but with
less XP and some reputation loss, so I feel this is fair balance-wise.