Do not reset music on map reload
Now the ambient music doesn't reset if the track is the same!  Finally,
I can listen them to the end even while quickloading every 15 seconds.
This is best used with infinite music loop option from the patch,
without it restarting the track is a tad more difficult than quickload,
requiring changing the map, loading another save, or exiting to the main
screen (and the first two won't work if the track isn't different).