Add some fae dust to Hall under the Hill
Since yellow reagents are now more important early on (for magic
potions), I'll have to add some sources for low-level ones, which are
lacking in vanilla.  For this commit, I've spread 20 bottles of fae dust
in the Avlee fairy hall.  This will break 3.0 games, so I'll probably
have to make a separate fae-dust-free compatibility release.