Let gogs drop sulfur
Continuing the proliferation of yellow reagents, gogs of all types now
have a small chance to be looted for sulfur.  It's specifically looted
from a corpse, unlike the usual bonus reagent drops, because I wanted
the chance to be lower than 20%, and also because I don't like monsters
leaving useless corpses (with no money or items).  But most of corpse
changes will wait until 4.0, as I don't want to break saves too much.