Fix repeated Armageddon displacing flyers into sky
Now that Armageddon can be cast as much as ten times per day, it's more
of an issue that it raises flying monsters into the sky, like all
others, except the flyers don't fall down afterwards.  With enough
castings, it's possible to displace the monsters high enough to be
beyond party reach.  This is stupid, so this commit prevents flying
monsters from being moved upwards during the Armageddon sequence.

Also, I've noticed during testing that Elixir of Life sometimes fails to
heal unconscious PCs due to a typo, so that's now fixed, too.