Add spell-absorbing robe enchantment
It irked me a little that the new Master Leather perk, while being mage-
oriented, is not extended to robes which are also supposed to be for
mages.  So, in lieu of a skill perk, as robes have no skill, they may
now have a 10% spell absorption chance as a rare enchantment.