Fix XP-boosting NPCs granting a 9% bonus w/o skill
A tiny vanilla bug: with a teacher or a similar NPC hired, all PCs were
considered to have the Learning skill, and thus gained the 9% bonus to
XP that the skill grants.  Although it's a good idea to learn this skill
early anyway, it's still wrong to give the bonus without it.  Fixed.

Also here: apparently Scholar's Russian dismissal text was cut off.