Fix Telekinesis interacting with a wrong PC
Turns out that aiming a spell in realtime mode imposes 300 recovery
(approx. 140 when accounting for the realtime slowdown factor), which
also switches the current PC, which causes items or corpses or chests to
be opened or picked up by another PC than which cast the Telekinesis
spell.  It's what I wanted to fix, but that 300 (140) recovery was also
bad, as most aimed spells have less than 140, and it was effectively
raised to that value.  Quick spells weren't affected, though.

Also here: rewrite some old code as I now know what struct it uses.

The first version of this commit also removed the 32/15th realtime
recovery slowdown, but it turned out to make realtime combat way too
frantic, especially if you just want to run past the monsters.