Allow loading 3.0 saves for 3.1
Out of concern for both of the people who already started playing v3.0
of the mod, this commit makes sure that the saves can be loaded in 3.1.
Obviously, this won't retroactively e.g. insert new item enchantments
onto existing items, but I did at least write the code that adds sulfur
to present gogs (including corpses).  The game also apparently
force-resets Hall under the Hill which was changed after 3.0, so this
saves me some effort (not sure how reliable this mechanic is, though!)