Add prefix flag to spcitems.txt
For the next change, I need a better way to figure out if an enchantment
name is a prefix or postfix -- in vanilla it's hardcoded as many cmps,
and until now I just added more.  But I'd rather have an array, so why
not add it as a spcitems field -- this is prettier, and as an added
bonus I can change prefix status between English and Russian versions!

In fact, I immediately moved several prefixes that didn't translate well
to postfix status in Russian.  And while I was at it, I renamed some
enchantment names to eliminate duplicates -- this will also be helpful.