Add shop orders, part 3
Now the parsed item will be displayed on a order preview screen, along
with (fixed for now) ore, money, and a special reagent.  I still need to
decide how much of all that to charge for which items, and to obtain a
proper blueprint background somehow, but we're getting there.

Also here: removed E-dots from some spc bonus names for easier parsing,
and added order prompts to armor and magic shops that I forgot before.