Add shop orders, part 4
Now we have the code to compute gold, ore, reagent, and time.
All that's left is to actually allow ordering the item, and to get
graphics for order screen background and the order button.

Also here: Elixir of Life scroll and book titles weren't updated;
and the gender-specific artifact Russian text needed updating too.