Fix auto-ID of items in chests
Turns out my 3.0 code that auto-identified static zero-difficulty items
in chests (e.g. herbs in Welnin stable crates) also randomly identified
other items in the same chest.  I've moved the hook to the routine that
is called when a chest is first opened, and it seems to work fine now.