Implement the new Sniper quest
Now light and dark Archer promotions no longer have essentially the same
quest, as Sniper are required to fetch a magic quiver from the new genie
location instead.  The quiver is a static artifact with two bonuses, and
the Perfect Bow is now too, to balance the rewards between paths.  I did
not add new graphics for the quiver yet, though, that can wait a bit.

Also here: added a unique key for genie map, and swapped duplicate key
images with unused ones while I was at it; sequestered the code for crit
chance, which is now called twice instead of duplicate assembly logic;
and Kaltenberg updated Bracada minimap for me as I forgot about that.