Allow launching spells in arbitrary directions
A long-awaited QoL feature particularly important for the genie map: now
it's possible to shoot projectile spells in the direction of the cursor,
even with no target there, by shift-clicking while aiming the spell.
Mostly it's useful for AOE spells like Fireball or Rock Blast -- the
latter is much more convenient to bounce now, and I removed the special
code causing it to always shoot forward as I believe it was NWC's clumsy
way of allowing it to target walls, and we have a better way now.