Make meteor spells aimed and tweak Death Blossom
Now Meteor Shower and Starburst can be aimed too, and Shift-clicking
will target them in the horisontal direction of the cursor, with its
vertical position corresponding to distance from party instead.

Death Blossom is also aimed now.  With a valid monster target, it will
calculate launch angle so that it flies in an arc directly at the
monster.  This actually changes the usual effect from the spell, as the
target is no longer bombarded from above, but at least it hits flying
monsters now, with some of the shards anyway.  In any case, it can also
be manually aimed with Shift, with the rock flying in the cursor
direction, but with a 45 degree bonus to the verical angle.

Also it turned out that there was no actual bonus to the latter spell at
GM, so now it doubles the shard count, which is 8 at Master (up from 7).