Boost Conflict Aura duration and allow removing it
Aura of Conflict had only 10 minutes per level duration in v3, mostly
because I was afraid the player would be stuck with a handicap buff if
the tactical situation changes before it expires.  This was ultimately
the wrong approach, since things can change very quickly anyway, besides
I already have a working Ctrl-click-to-cancel-spell mechanism that could
be used here as well.  So it is now, and I'm free to increase the
duration to something that makes the spell practical even early on: one
hour, plus 15 minutes per skill point.  Thus it starts at 75 minutes,
which is already comfortably long (for an early party, anyway).

Also here: tweaked some English spell descriptions a bit.