Add a hotkey for drinking a healing potion
Idea from MAW.  This commands searches for either the weakest potion
that will heal the current PC fully, or failing that, the strongest
available, looking in all four backpacks.  The PC must be active and not
recovering, and the action has the usual 100 recovery delay.  I haven't
assigned any key to it by default, since not everyone will want that.