Rework the beacon penalty to mana heal
I want to penalize using Lloyd's Beacon just to replenish the SP pool.
The first attempt was to increase the price of guild SP restoration
after recalling a beacon, but it doesn't even work since one typically
heals, THEN teleports back.  Now that I can add PC buffs, it can be done
better: using a guild adds a two-hour "buff" which will remove the
restored SP and damage the PC if a beacon is recalled while it's still
active.  I suppose this can still be circumvented by letting an enemy
Dispel it, but if one is willing to go that far, they deserve it.