Make repair take a turn
Now using the Repair skill imposes 100 recovery, and it can only be done
by the currently active PC, same as a scroll or potion.  Mostly this is
just for realism, since mid-combat repair never quite made sense to me.

Notably, smith hirelings are an exception for this rule and will repair
anything instantly -- hopefully this will make them a bit less useless.