Adjust Hero's Belt graphics and effects
One, now it has unique gfx, from MM6.  Since it's male-only, there was
conveniently no need for two equipped sprites, which MM6 does not have.

Two, it no longer increases resistances.  I now believe it was a bug, an
unintentional fallthrough to the Lady's Escort code, which explains why
it wasn't mentioned in vanilla description; it was too much anyway.