Add a consumable that forces map refill
For now it can only be obtained as an Arena reward (for every 5 "victory
points").  The gold reward is reduced from x5 to x3 as a compensation.

Using it indoors will reset the map on next visit, outdoors it only
advances the counter by half of the refill period, so two may be needed.

Also, entering any map will print a statusline hint if it's about to
reset (in less than a month), or if it already reset just now.

I will need to do something to prevent duplicating unique items later.

Also here: changed unconsciousness monster bonus display from "stun" to
"knockout", because we already have a Stun spell, and it was confusing.